RabbitMQ + celery/rabbitMQWithLocalApp/rabbitmq/pyproject.toml
[project]
name = "rabbitmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
